Port to GTK4
The application has been ported to GTK4. This has required many trivial changes
(e.g., replacement of gtk_box_pack_start() with gtk_box_append()), and a few
substantial changes (replacement of code pathway for coloring symbolic icons,
use of a GtkPopover for the WPS menu instead of a GtkMenuButton, and use of
GtkEventController for key bindings instead of g_signal_connect()).