Use detailed_errors_generic[] as fallback
The error code lookup logic in validation_callback() has been changed. The error
table in the CallbackMessages struct is now optional. If a custom error table is
supplied, it will be checked first. If a custom table is not supplied, or if no
suitable error message is found, validation_callback() is now hard-coded to fall
back to detailed_errors_generic[].