Hold indefinitely when run with -i option
Formerly, the application was only held open for as long as indicator icons were
actually present. For example, if iwd were stopped, or if all wireless devices
were physically detached, iwgtk would close. Now iwgtk is held open
indefinitely, regardless of whether any indicators are active, when run with the
-i option.