Prevent UI widgets from moving and resizing
Set some minimum widget sizes, to prevent widgets from moving and resizing in
response to mode changes and connection/disconnection events. This solution is
only a proof of concept: the minimum sizes are hard-coded, and the values chosen
will not work for different themes or translations. A more robust solution would
require that minimum widget sizes be dynamically calculated.