Use Meson build system
In preparation for internationalization via gettext, Meson has been adopted as
the project's build system. The old Makefile has been removed, and the
installation instructions in the README have been updated. The installation
directory for iwgtk.conf is now determined by Meson's sysconfdir parameter;
previously, it was hard-coded to /etc in the Makefile.